Extend print-method to print SystemMap tersely
System objects tend to be large and unweildy, with a lot of duplicated
elements. Printing them at the REPL is rarely useful. Some systems
contain large objects, such as byte arrays, that can lock up a REPL
for several minutes when you try to print them.

This change extends clojure.core/print-method to SystemMap so that it
prints only a minimal indication of its type.

I considered also printing a list of keys in the system, but even that
could be large and it would be unlike any other printed representation
in Clojure.

As a consequence, SystemMap instances are no longer `read`able.

This behavior can be disabled with `remove-method`.