More specific error for components which are nil
Before this change, a nil value in a system map would cause an error
::missing-component.

However, the most common cause of this error is a component which failed
to return a value from `start` or `stop`. We can distinguish this error
case from a component whose key was not in the system at all, and return
a more informative error message.

See also https://github.com/stuartsierra/component/issues/17