Implement AutoCloseable on SystemMap
This allows a SystemMap to be used in a `with-open` expression,
with `.close` as an alias for `stop`.

This could be either java.lang.AutoCloseable or java.io.Closeable.
Closeable is older, introduced in Java 5,
but it is defined specifically for IO streams:
The JavaDoc says it "Closes this stream,"
and can throw IOException.
AutoCloseable is newer, introduced in Java 7,
and more general.
Its JavaDocs says it "Closes this resource,
relinquishing any underlying resources,"
and it throws Exception.

Nearly all versions of Clojure require Java 7+,
and I beleive recent versions require Java 8+,
so depending on AutoCloseable should be safe.

I am not sure how to make this compatible with ClojureScript.
I cannot find an equivalent to "AutoCloseable" or "Closeable"
defined in ClojureScript, and I am not aware of a JavaScript
equivalent either.

Perhaps I could define a private "Closeable" protocol in
com.stuartsierra.component.platform for compatibility.