Type hint system-map return value
The `with-open` calls `.close` on its argument, without making any
assumptions about how the compiler will resolve the .close method.
This can lead to reflection warnings if the type is not known.

This might lead to a lot of clutter if I also have to type-hint all of
the functions that work with system-maps, such as `.close`. It might
not even make sense. Maybe I should just leave it up to the user to
type-hint the system as AutoCloseable if they care about reflection
warnings.