Revert "Type hint system-map return value"
This reverts commit 033273df597bf62f9610bafe1e87db2ae88ca533.

In a typical usage, for example in a test,
I expect this to be used like this:

    (with-open [system (start (system-map ...))]
      ...)

I cannot type-hint the return value of `start`,
because it is a protocol method that might be implemented
on components that do *not* implement AutoCloseable.

So there is not much point in type-hinting the constructor.