Implement java.io.Closeable on SystemMap
java.lang.AutoCloseable ("Closes this resource") would be more
technically correct than java.io.Closeable ("Closes this stream"). But
Closeable is more well-known. If I see a reflection warning from
`with-open`, I will likely add a type hint for `^Closeable` instead of
`^AutoCloseable`.

If someone does type hint a SystemMap as `^AutoCloseable`, it will
still work, because Closeable is a sub-interface of AutoCloseable.

This implementation is restricted to Clojure (JVM) only, with no
ClojureScript alternative. Based on my casual research, there is no
well-known standard for "Closeable" in JavaScript. There is a proposal
for "Explicit Resource Management" but it does not appear to be
universally-supported yet:

https://github.com/tc39/proposal-explicit-resource-management

https://v8.dev/features/explicit-resource-management

This allows `with-open` to be used for tests
instead of adding a custom "with-*" macro as proposed in
https://github.com/stuartsierra/component/issues/6 and
https://github.com/stuartsierra/component/issues/7